only copy and clean if build succeeded

+++ builder_job.sh
@@ -31,10 +31,14 @@ cp builder.sh /scratch/$USER/builder
 podman run --rm --network=host -v /scratch/$USER/builder:/root public.ecr.aws/ubuntu/ubuntu:latest /root/builder.sh
 
 cd /scratch/$USER/builder/android/pe/out/target/product/walleye
-build_id="$(basename PixelExperience_*.zip .zip)"
-mkdir -p ~/builds/"$build_id"
-cp PixelExperience_* ~/builds/"$build_id"
-cp boot.img ~/builds/"$build_id"
-cd
+shopt -s nullglob
+zip_file="$(echo PixelExperience_*.zip)"
+if [[ -f "$zip_file" ]]; then
+    build_id="$(basename "$zip_file" .zip)"
+    mkdir -p ~/builds/"$build_id"
+    cp PixelExperience_* ~/builds/"$build_id"
+    cp boot.img ~/builds/"$build_id"
+    cd
 
-podman unshare rm -rf /scratch/$USER/builder
+    podman unshare rm -rf /scratch/$USER/builder
+fi
